requirements.txt
openai
pdfplumber
tiktoken
gspread
python-dotenv
oauth2client